requirements.txt
cachelib==0.13.0
Flask==3.1.2
Flask_SocketIO==5.5.1
GitPython==3.1.46
Jinja2==3.1.6
tinydb==4.8.2
